requirements.txt
Django~=2.2.4
djangorestframework~=3.11.1
coverage